Removing yarnrc requirement
The yarn cli issue with --no-bin-links has been fixed, and is now
correctly recognized when passed directly on the command line. As such,
removing the autogenerated yarnrc and instead injecting --no-bin-links
and --modules-folder on actual yarn command invocation.

As a side-effect, node_modules no longer needs to be a subdirectory of
the current working directory for yarn / node to work properly.